Add Game Exiting Logic (WIP)
